repeater-book/requirements.txt
attrs
requests
pyquery
